uint32 time

+++ utils/cLog.h
@@ -19,5 +19,5 @@ public:
   static void log (enum eLogCode logCode, std::string logStr);
   static void log (enum eLogCode logCode, const char *format, ... );
 
-  static std::string getLine (int n, eLogCode& logCode, uint64_t& usTime);
+  static std::string getLine (int n, eLogCode& logCode, uint32_t& usTime);
   };
